update: change seat layout button color and position of go home button
